Modernize I/O
1. Use Popen.communicate
Some apps have huge output and the code would hang.
For example for `python3 adbe.py app-info com.google.android.gms`

2. Pipe directly
The fear I had and still have while piping directly is that if there is
an error in the first adb command, the second command might end up
ignoring it but as I tested repeatedly that's not happening.
This also eliminates a code branch and now we have a single execution
code path.