Fix escaping
Many commands like screenshot, ls which use _may_be_wrap_with_run_as
method were broken. This diff fixes the shell escaping. I am not
fully-confident if this works but this seems to work in most of the
tests I performed on an emulator (API 16) and a real device (API 28)