Cleanup run-as and root
The approach to handle file related commands was clumsy. I was stiching
a single command with both "run-as" and "root" prefixes creating three
versions which were combined and passed to adb shell as a single
command. This required convluted escaping as well.

I replaced that with a simpler approach where the caller passes a
filepath to a single method which decides whether to run this command
with run-as or even try root. It will then run the three versions one
by one and stops on the first success. This eliminates a lot of clumsy
logging as well.

Also, fixed a big in push file where the temporary file was being
deleted before being copied to its final location.